+++ lua/obsidian/commands/rename.lua
@@ -1,9 +1,8 @@
 ---@param data obsidian.CommandArgs
 return function(data)
-  local new_name = vim.trim(data.args)
-  if #new_name == 0 then
-    vim.lsp.buf.rename()
-  else
-    vim.lsp.buf.rename(new_name)
+  local new_name
+  if data.args and string.len(data.args) > 0 then
+    new_name = vim.trim(data.args)
   end
+  vim.lsp.buf.rename(new_name)
 end